[WFLY-13848] Move the "org.jboss.as.security.plugins" package into it's own module.
