+++ ee-feature-pack/common/src/main/resources/modules/system/layers/base/org/jboss/as/webservices/main/module.xml
@@ -57,6 +57,7 @@
         <module name="org.jboss.as.naming"/>
         <module name="org.jboss.as.network"/>
         <module name="org.jboss.as.security"/>
+        <module name="org.jboss.as.security-plugins"/>
         <module name="org.wildfly.security.elytron-private"/>
         <module name="org.jboss.as.web-common"/>
         <module name="org.jboss.modules"/>
