+++ ee-feature-pack/common/src/main/resources/modules/system/layers/base/org/wildfly/extension/messaging-activemq/main/module.xml
@@ -66,6 +66,7 @@
         <module name="org.jboss.ironjacamar.api"/>
         <module name="org.jboss.as.connector" />
         <module name="org.jboss.as.security"/>
+        <module name="org.jboss.as.security-plugins"/>
         <module name="org.jboss.vfs"/>
         <module name="org.jboss.jts"/>
         <module name="org.picketbox"/>
