+++ security/plugins/src/main/java/org/jboss/as/security/lru/RemoveCallback.java
