+++ security/plugins/src/main/java/org/jboss/as/security/plugins/AuthenticationCacheFactory.java
