+++ security/plugins/src/main/java/org/jboss/as/security/plugins/DefaultAuthenticationCacheFactory.java
