+++ security/pom.xml
@@ -44,6 +44,7 @@
     <modules>
         <module>api</module>
         <module>integration</module>
+        <module>plugins</module>        
         <module>subsystem</module>
     </modules>
