+++ security/subsystem/pom.xml
@@ -102,6 +102,12 @@
 
         <dependency>
             <groupId>${project.groupId}</groupId>
+            <artifactId>wildfly-security-plugins</artifactId>
+            <scope>provided</scope>
+        </dependency>
+
+        <dependency>
+            <groupId>${project.groupId}</groupId>
             <artifactId>wildfly-transactions</artifactId>
         </dependency>
 
@@ -139,13 +145,13 @@
         </dependency>
 
         <dependency>
-        	<groupId>org.picketbox</groupId>
-        	<artifactId>picketbox</artifactId>
+            <groupId>org.picketbox</groupId>
+            <artifactId>picketbox</artifactId>
         </dependency>
 
         <dependency>
-        	<groupId>org.jboss.spec.javax.security.jacc</groupId>
-        	<artifactId>jboss-jacc-api_1.5_spec</artifactId>
+            <groupId>org.jboss.spec.javax.security.jacc</groupId>
+            <artifactId>jboss-jacc-api_1.5_spec</artifactId>
         </dependency>
 
         <dependency>
