+++ servlet-feature-pack/common/src/main/resources/modules/system/layers/base/org/jboss/as/security-plugins/main/module.xml
@@ -2,7 +2,7 @@
 
 <!--
   ~ JBoss, Home of Professional Open Source.
-  ~ Copyright 2010, Red Hat, Inc., and individual contributors
+  ~ Copyright 2020, Red Hat, Inc., and individual contributors
   ~ as indicated by the @author tags. See the copyright.txt file in the
   ~ distribution for a full listing of individual contributors.
   ~
@@ -22,49 +22,23 @@
   ~ 02110-1301 USA, or see the FSF site: http://www.fsf.org.
   -->
 
-<module xmlns="urn:jboss:module:1.5" name="org.jboss.as.security">
+<module xmlns="urn:jboss:module:1.5" name="org.jboss.as.security-plugins">
     <properties>
         <property name="jboss.api" value="private"/>
     </properties>
 
-    <exports>
-        <exclude path="org/jboss/as/security/logging"/>
-    </exports>
-
     <resources>
-        <artifact name="${org.wildfly:wildfly-security}"/>
+        <artifact name="${org.wildfly:wildfly-security-plugins}"/>
     </resources>
 
     <dependencies>
-        <module name="javax.api"/>
         <module name="javax.security.jacc.api"/>
-        <module name="javax.servlet.api"/>
-        <module name="javax.transaction.api"/>
-        <module name="org.apache.commons.cli" optional="true"/>
-        <module name="org.infinispan" optional="true"/>
-        <module name="org.infinispan.commons" optional="true"/>
-        <module name="org.jboss.as.core-security"/>
-        <module name="org.jboss.as.controller"/>
-        <module name="org.jboss.as.domain-management"/>
-        <module name="org.jboss.as.ee"/>
-        <module name="org.jboss.as.server"/>
-        <module name="org.jboss.as.security-plugins"/>
-        <module name="org.jboss.as.naming"/>
         <module name="org.jboss.logging"/>
-        <module name="org.jboss.metadata.common"/>
-        <module name="org.jboss.metadata.web"/>
-        <module name="org.jboss.metadata.ear"/>
         <module name="org.jboss.modules"/>
-        <module name="org.jboss.msc"/>
-        <module name="org.jboss.remoting"/>
-        <module name="org.jboss.security.negotiation" optional="true"/>
-        <module name="org.jboss.staxmapper"/>
         <module name="org.picketbox"/>
-        <module name="org.wildfly.common"/>
         <module name="org.wildfly.clustering.infinispan.spi" optional="true"/>
-        <module name="org.wildfly.clustering.service" optional="true"/>
+        <module name="org.wildfly.common"/>
         <module name="org.wildfly.security.elytron-private"/>
-        <module name="org.wildfly.extension.undertow" />
         <module name="sun.jdk"/>
     </dependencies>
 </module>
